vhost-device-vsock: Move test utils to test_utils module
Move HeadParams and prepare_desc_chain_vsock from vsock_conn tests
into a shared test_utils module so they can be reused by
thread_backend tests.
