+++ vhost-device-vsock/src/main.rs
@@ -45,6 +45,9 @@ mod vhu_vsock;
 mod vhu_vsock_thread;
 mod vsock_conn;
 
+#[cfg(test)]
+mod test_utils;
+
 use std::{
     any::Any,
     collections::HashMap,
